+++ src/qudi/logic/scanning_data_logic.py
@@ -287,7 +287,7 @@ class ScanningDataLogic(LogicBase):
                         arrowprops={'facecolor': '#17becf', 'shrink': 0.05})
         return fig
 
-    def save_scan(self, scan_data, color_range=None):
+    def save_scan(self, scan_data, color_range=None, custom_tag=None):
         with self._thread_lock:
             if self.module_state() != 'idle':
                 self.log.error('Unable to save 2D scan. Saving still in progress...')
@@ -335,7 +335,10 @@ class ScanningDataLogic(LogicBase):
                 for channel, data in scan_data.data.items():
                     # data
                     # nametag = '{0}_{1}{2}_image_scan'.format(channel, *scan_data.scan_axes)
-                    tag = self.create_tag_from_scan_data(scan_data, channel)
+                    if custom_tag:
+                        tag = custom_tag
+                    else:
+                        tag = self.create_tag_from_scan_data(scan_data, channel)
                     file_path, _, _ = ds.save_data(data,
                                                    metadata=parameters,
                                                    nametag=tag,
@@ -356,10 +359,10 @@ class ScanningDataLogic(LogicBase):
                 self.sigSaveStateChanged.emit(False)
             return
 
-    def save_scan_by_axis(self, scan_axes=None, color_range=None):
+    def save_scan_by_axis(self, scan_axes=None, color_range=None, custom_tag=None):
         # wrapper for self.save_scan. Avoids copying scan_data through QtSignals
         scan = self.get_current_scan_data(scan_axes=scan_axes)
-        self.save_scan(scan, color_range=color_range)
+        self.save_scan(scan, color_range=color_range, custom_tag=custom_tag)
 
     def create_tag_from_scan_data(self, scan_data, channel):
         axes = scan_data.scan_axes