+++ docs/changelog.md
@@ -14,6 +14,7 @@
 - Re-introduced tilt correction (from old core) to the scanning probe toolchain.
 - Improved support for Stanford Research Systems signal generators
 - Expanded documentation of the microwave interface
+- Added option to specify custom save file name in scanning GUI (PR #148)
 
 ### Bugfixes